Module Ppx_module_timer_runtime

val am_recording : Base.bool

If true, ppx_module_timer records module startup times and reports them on stdout at process exit. Controlled by am_recording_environment_variable.

val am_recording_environment_variable : Base.string

If this environment variable is set (to anything) when this module starts up, am_recording is set to true.

Equal to "PPX_MODULE_TIMER".

If this is set to a valid duration string (see Duration.format below), that duration is used to override recorded times for each module. This is used to make test output deterministic.

If this is set to "FAKE_MODULES", the entire set of recorded data is overridden with fake values. This is used to make test output both deterministic and stable, so that changes in external library dependencies do not affect it. The fake data is not particularly sensible, for example we are not careful to make the times for definitions add up to the time for the enclosing module.

module Duration : sig ... end